Move #more-words button to right on wide screens. Keep focus and hover synchronized for spelling area.
Moving #more-words away from spelling suggestions keeps mouse from hovering over a suggestion on the next set (depending on the lenghts of the various words) until the user moves mouse there. Keeping the focus where the mouse is and resetting the colors under the mouse when the focus is elsewhere makes it more clear what will happen when you press the Enter key.